+++ .github/workflows/test.yml
@@ -501,7 +501,12 @@ jobs:
             'eligibility/**/*.spec.js',
             'exporter/**/*.spec.js',
             'external/**/*.spec.js',
-            'facility/**/*.spec.js',
+            'facility/*.spec.js',
+            'facility/end-date/*.spec.js',
+            'facility/review-ukef-approval/*.spec.js',
+            'facility/version-0/*.spec.js',
+            'facility/unissued-facilities/AIN/*.spec.js',
+            'facility/unissued-facilities/MIN/*.spec.js',
             'maker/**/*.spec.js',
             'scenarios/**/*.spec.js',
             'security/**/*.spec.js',
@@ -733,6 +738,9 @@ jobs:
             './gef/cypress/e2e/exporter/**/*.spec.js',
             './gef/cypress/e2e/external/**/*.spec.js',
             './gef/cypress/e2e/facility/*.spec.js',
+            './gef/cypress/e2e/facility/end-date/*.spec.js',
+            './gef/cypress/e2e/facility/review-ukef-approval/*.spec.js',
+            './gef/cypress/e2e/facility/version-0/*.spec.js',
             './gef/cypress/e2e/facility/unissued-facilities/AIN/*.spec.js',
             './gef/cypress/e2e/facility/unissued-facilities/MIN/*.spec.js',
             './gef/cypress/e2e/maker/**/*.spec.js',
