+++ e2e-tests/gef/cypress/e2e/facility/end-date/facilities-change-facility-end-date-to-bank-review-date-journey.spec.js
@@ -1,9 +1,9 @@
-import relative from '../relativeURL';
-import aboutFacility from '../pages/about-facility';
-import bankReviewDate from '../pages/bank-review-date';
-import facilityEndDate from '../pages/facility-end-date';
-import { tomorrowDay, tomorrowMonth, tomorrowYear, todayDay, todayMonth, todayYear, tomorrow } from '../../../../e2e-fixtures/dateConstants';
-import { BANK1_MAKER1 } from '../../../../e2e-fixtures/portal-users.fixture';
+import relative from '../../relativeURL';
+import aboutFacility from '../../pages/about-facility';
+import bankReviewDate from '../../pages/bank-review-date';
+import facilityEndDate from '../../pages/facility-end-date';
+import { tomorrowDay, tomorrowMonth, tomorrowYear, todayDay, todayMonth, todayYear, tomorrow } from '../../../../../e2e-fixtures/dateConstants';
+import { BANK1_MAKER1 } from '../../../../../e2e-fixtures/portal-users.fixture';
 
 const applications = [];
 let token;
